Add configurable auto restart
The minimum number of players required can now be adjusted. The default value is defined in config.yml, and it can be overridden by using the max-players command argument.

The restart now sends notifications to both the game chat and the bridge channel and waits 600 ticks (30 seconds) before the restart occurs.

The server MOTD also indicates that the server is restarting and the player limit is set to zero.